[FIX] website_slides: move JS in the correct file
Commit https://github.com/odoo/odoo/commit/6f5748d28854305bb33ffe54e5f94bab4a14828f
split and moved JS in different files... but in files which were in
different assets.
