+++ addons/website_slides/views/assets.xml
@@ -3,9 +3,10 @@
     <data>
 
         <template id="assets_frontend" inherit_id="website.assets_frontend" name="Slide">
-            <xpath expr="." position="inside">
+            <xpath expr="//link[last()]" position="after">
                 <link rel="stylesheet" type="text/scss" href="/website_slides/static/src/scss/website_slides.scss" t-ignore="true"/>
-                <!-- Js widgets, ... order loading files is IMPORTANT -->
+            </xpath>
+            <xpath expr="//script[last()]" position="after">
                 <script type="text/javascript" src="/website_slides/static/src/js/slides.js"/>
                 <script type="text/javascript" src="/website_slides/static/src/js/slides_like.js"/>
                 <script type="text/javascript" src="/website_slides/static/src/js/slides_share.js"/>
